Fix type mismatch between value_hash prototype and definition.
Thanks to michel <michel@cecaps.ufmg.br> for pointing out the problem.

+++ src/data/value.h
@@ -72,7 +72,7 @@ void value_copy_buf_rpad (union value *dst, int dst_width,
 void value_set_missing (union value *, int width);
 int value_compare_3way (const union value *, const union value *, int width);
 bool value_equal (const union value *, const union value *, int width);
-size_t value_hash (const union value *, int width, unsigned int basis);
+unsigned int value_hash (const union value *, int width, unsigned int basis);
 
 bool value_is_resizable (const union value *, int old_width, int new_width);
 bool value_needs_resize (int old_width, int new_width);
